move Nexus's MGS-driven updates code into its own package
