+++ nexus/mgs-updates/tests/sp_updater.rs
@@ -9,9 +9,8 @@ use gateway_messages::{SpPort, UpdateInProgressStatus, UpdateStatus};
 use gateway_test_utils::setup as mgs_setup;
 use hubtools::RawHubrisArchive;
 use hubtools::{CabooseBuilder, HubrisArchiveBuilder};
-use omicron_nexus::app::test_interfaces::{
-    MgsClients, SpUpdater, UpdateProgress,
-};
+use nexus_mgs_updates::{MgsClients, SpUpdater, UpdateProgress};
+use slog::debug;
 use sp_sim::SIM_GIMLET_BOARD;
 use sp_sim::SIM_SIDECAR_BOARD;
 use sp_sim::{SimSpHandledRequest, SimulatedSp};
