+++ nexus/src/app/update.rs
@@ -16,27 +16,6 @@ use omicron_common::api::external::{
 use semver::Version;
 use update_common::artifacts::{ArtifactsWithPlan, ControlPlaneZonesMode};
 
-mod common_sp_update;
-mod host_phase1_updater;
-mod mgs_clients;
-mod rot_updater;
-mod sp_updater;
-
-pub use common_sp_update::SpComponentUpdateError;
-pub use host_phase1_updater::HostPhase1Updater;
-pub use mgs_clients::MgsClients;
-pub use rot_updater::RotUpdater;
-pub use sp_updater::SpUpdater;
-
-#[derive(Debug, PartialEq, Clone)]
-pub enum UpdateProgress {
-    Started,
-    Preparing { progress: Option<f64> },
-    InProgress { progress: Option<f64> },
-    Complete,
-    Failed(String),
-}
-
 impl super::Nexus {
     pub(crate) async fn updates_put_repository(
         &self,
