+++ nexus/tests/integration_tests/mod.rs
@@ -21,7 +21,6 @@ mod demo_saga;
 mod device_auth;
 mod disks;
 mod external_ips;
-mod host_phase1_updater;
 mod images;
 mod initialization;
 mod instances;
@@ -38,7 +37,6 @@ mod quotas;
 mod rack;
 mod role_assignments;
 mod roles_builtin;
-mod rot_updater;
 mod router_routes;
 mod saml;
 mod schema;
@@ -46,7 +44,6 @@ mod silo_users;
 mod silos;
 mod sleds;
 mod snapshots;
-mod sp_updater;
 mod ssh_keys;
 mod subnet_allocation;
 mod support_bundles;
